requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.5.2
redis==5.0.3
pydantic==2.6.4
python-dotenv==1.0.1
pypdf2==3.0.1
python-multipart==0.0.9
PyYAML==6.0.1